Move SDLXX_CHECK to new header macros.hpp
